Check for existence of setup.py before trying to install

+++ readthedocs/projects/tasks.py
@@ -356,15 +356,18 @@ def update_imported_docs(version_pk):
                         cmd=project.venv_bin(version=version_slug, bin='pip'),
                         requirements=project.requirements_file))
             os.chdir(project.checkout_path(version_slug))
-            if getattr(settings, 'USE_PIP_INSTALL', False):
-                update_docs_output['install'] = run(
-                    '{cmd} install --force-reinstall --ignore-installed .'.format(
-                        cmd=project.venv_bin(version=version_slug, bin='pip')))
+            if os.path.isfile("setup.py"):
+                if getattr(settings, 'USE_PIP_INSTALL', False):
+                    update_docs_output['install'] = run(
+                        '{cmd} install --force-reinstall --ignore-installed .'.format(
+                            cmd=project.venv_bin(version=version_slug, bin='pip')))
+                else:
+                    update_docs_output['install'] = run(
+                        '{cmd} setup.py install --force'.format(
+                            cmd=project.venv_bin(version=version_slug,
+                                                 bin='python')))
             else:
-                update_docs_output['install'] = run(
-                    '{cmd} setup.py install --force'.format(
-                        cmd=project.venv_bin(version=version_slug,
-                                             bin='python')))
+                update_docs_output['install'] = (999, "", "No setup.py, skipping install")
 
         # Update tags/version
